[change] Move UserAccounting back to Views
