+++ openwisp_radius/api/views.py
@@ -1,6 +1,7 @@
 import logging
 from uuid import UUID
 
+import swapper
 from dj_rest_auth import app_settings as rest_auth_settings
 from dj_rest_auth.registration.views import RegisterView as BaseRegisterView
 from dj_rest_auth.views import PasswordChangeView as BasePasswordChangeView
@@ -20,13 +21,19 @@ from django.utils.encoding import force_bytes, force_text
 from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
 from django.utils.translation import gettext_lazy as _
 from django.views.decorators.csrf import csrf_exempt
+from django_filters.rest_framework import DjangoFilterBackend
 from drf_yasg.utils import no_body, swagger_auto_schema
 from rest_framework import serializers, status
 from rest_framework.authentication import SessionAuthentication
 from rest_framework.authtoken.models import Token as UserToken
 from rest_framework.authtoken.views import ObtainAuthToken as BaseObtainAuthToken
 from rest_framework.exceptions import AuthenticationFailed, NotFound, ParseError
-from rest_framework.generics import CreateAPIView, GenericAPIView, RetrieveAPIView
+from rest_framework.generics import (
+    CreateAPIView,
+    GenericAPIView,
+    ListAPIView,
+    RetrieveAPIView,
+)
 from rest_framework.permissions import (
     DjangoModelPermissions,
     IsAdminUser,
@@ -42,30 +49,36 @@ from .. import settings as app_settings
 from ..exceptions import PhoneTokenException
 from ..utils import generate_pdf, load_model
 from . import freeradius_views
+from .freeradius_views import AccountingFilter, AccountingViewPagination
 from .permissions import IsRegistrationEnabled, IsSmsVerificationEnabled
 from .serializers import (
     AuthTokenSerializer,
     ChangePhoneNumberSerializer,
+    RadiusAccountingSerializer,
     RadiusBatchSerializer,
     ValidatePhoneTokenSerializer,
 )
 from .swagger import ObtainTokenRequest, ObtainTokenResponse, RegisterResponse
-from .utils import DispatchOrgMixin, ErrorDictMixin, ThrottledAPIMixin
+from .utils import ErrorDictMixin
 
 authorize = freeradius_views.authorize
 postauth = freeradius_views.postauth
 accounting = freeradius_views.accounting
-user_accounting = freeradius_views.user_accounting
 
 _TOKEN_AUTH_FAILED = _('Token authentication failed')
 renew_required = app_settings.DISPOSABLE_RADIUS_USER_TOKEN
 logger = logging.getLogger(__name__)
 
 User = get_user_model()
+Organization = swapper.load_model('openwisp_users', 'Organization')
 PhoneToken = load_model('PhoneToken')
 RadiusAccounting = load_model('RadiusAccounting')
 RadiusToken = load_model('RadiusToken')
 RadiusBatch = load_model('RadiusBatch')
+
+
+class ThrottledAPIMixin(object):
+    throttle_scope = 'others'
 
 
 class BatchView(ThrottledAPIMixin, CreateAPIView):
@@ -99,6 +112,26 @@ class BatchView(ThrottledAPIMixin, CreateAPIView):
 
 
 batch = BatchView.as_view()
+
+
+class DispatchOrgMixin(object):
+    def dispatch(self, *args, **kwargs):
+        try:
+            self.organization = Organization.objects.select_related(
+                'radius_settings'
+            ).get(slug=kwargs['slug'])
+        except Organization.DoesNotExist:
+            raise Http404('No Organization matches the given query.')
+        return super().dispatch(*args, **kwargs)
+
+    def validate_membership(self, user):
+        if not (user.is_superuser or user.is_member(self.organization)):
+            message = _(
+                f'User {user.username} is not member of '
+                f'organization {self.organization.slug}.'
+            )
+            logger.warning(message)
+            raise serializers.ValidationError({'non_field_errors': [message]})
 
 
 class DownloadRadiusBatchPdfView(ThrottledAPIMixin, DispatchOrgMixin, RetrieveAPIView):
@@ -310,6 +343,49 @@ class ValidateAuthTokenView(DispatchOrgMixin, RadiusTokenMixin, CreateAPIView):
 
 
 validate_auth_token = ValidateAuthTokenView.as_view()
+
+
+class UserAccountingFilter(AccountingFilter):
+    class Meta(AccountingFilter.Meta):
+        fields = [
+            field for field in AccountingFilter.Meta.fields if field != 'username'
+        ]
+
+
+@method_decorator(
+    name='get',
+    decorator=swagger_auto_schema(
+        operation_description="""
+        **Requires the user auth token (Bearer Token).**
+        Returns the radius sessions of the logged-in user and the organization
+        specified in the URL.
+        """,
+    ),
+)
+class UserAccountingView(ThrottledAPIMixin, DispatchOrgMixin, ListAPIView):
+    authentication_classes = (BearerAuthentication, SessionAuthentication)
+    permission_classes = (IsAuthenticated,)
+    serializer_class = RadiusAccountingSerializer
+    pagination_class = AccountingViewPagination
+    filter_backends = (DjangoFilterBackend,)
+    filter_class = UserAccountingFilter
+    queryset = RadiusAccounting.objects.all().order_by('-start_time')
+
+    def list(self, request, *args, **kwargs):
+        self.request = request
+        return super().list(request, *args, **kwargs)
+
+    def get_queryset(self):
+        if getattr(self, 'swagger_fake_view', False):
+            return super().get_queryset()  # pragma: no cover
+        return (
+            super()
+            .get_queryset()
+            .filter(organization=self.organization, username=self.request.user.username)
+        )
+
+
+user_accounting = UserAccountingView.as_view()
 
 
 class PasswordChangeView(ThrottledAPIMixin, DispatchOrgMixin, BasePasswordChangeView):
